% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA25168238-package.R
\docType{package}
\name{SA25168238-package}
\alias{SA25168238}
\alias{SA25168238-package}
\title{SA25168238 package}
\description{
This package collects homework answers as vignettes and provides two research functions implemented in R and Rcpp. The package includes a permutation Cramer-von Mises test and a random-walk Metropolis sampler.
}
\author{
\strong{Maintainer}: wang wei \email{1484543635@qq.com}

}
\keyword{internal}
